Add Support for Refresh for I1 Devices
Refresh as an i1 device will only occur if the device engine is
known to be i1.  Otherwise default is to use standard i2/i2cs
type commands.

Scanning a db using standard length messages is a bit messy, so
this requires adding a DeviceScanManagerI1 to manage and send all
of the necessary requests.

Once a full db record is received it is passed to DeviceEntryI1
which inherits almost all of its functions from DeviceEntry,
however, the format of the record is slightly different.