Enable Writing I1 DB When New Address is Needed
This concept is very weird.  I don't think we have ever considered
this issue in any other Insteon protocol I have worked on.

It may be possible that when setting the last or highwater mark
that the same flag is set for all subsequent records.  However, I
am not sure if this is the case and don't have the patience to test
this out, so being cautious is probably best.