Move Extended Message CRC_Type to Constructor
The selection of CRC_Type was buried in to_bytes() which is only
called at the protocol level when sending a message.  Making it
unhelpful.

I moved the CRC_Type to the constructors so that it can be
designated at the time the message is created.  This is necessary
for adding thermostat functionality.