Add Join Command a Batch Linking Command
Mostly a command sequence.  Checks to see what the engine version
is, since linking will not work on i1 devices and is not required
for i2 devices.

If the device is i2cs, creates the modem->device link which is
the only one required for the modem to talk to the device.

Join is an idempotent (what a word!) function.  It will not
alter the state of the device more than necessary if run
multiple times.

I chose the word JOIN as I think it gives the right connotation
to the user even though it isn't an insteon term.  But this way
it won't be confused with a link or pair operation.  So new users
can quickly be instructed to join without leading to confusion.