Add Get Model Command; Run on Refresh if not Known
Get Model command will query the device for its device category,
sub-category, and firmware version.

This information is usually sufficient to determine the device
model and its full set of capabilities.

A Get Model command will run on refresh if:
1. The model number is not currently known; OR
2. Force on refresh is true

The Get Model command can also be run independently via mqtt
or the command line.

Adds testing for the function and documentation as well.