Ignore Links Created by Join or Pair Commands
Will not remove links associated with join or pair.  Also won't
add them.

We somewhat blindly handle pair, since we don't have any list of
available groups for each device.  This could be fixed by adding
in a static list to each object if necessary.

Had to change passed object to the device rather than the db.
This enables us to use methods on the device, such as find.