Add Import Scenes Function
Only added to devices currently, and does not save results to disk.

This code is complicated and rather spaghetti like.  This is in
part because we allow a huge amount of latitude for the users
to specify scenes.

This process is undoubtedly complex, but the concept is that
users will rarely use this feature.