Remove _Data Structure from SceneDevice; Fully Enable Pretty Names
All of SceneDevice data is now handled entirely in _yaml_data.
There is no longer a clean object _data that tracks the values.
This is almost certainly safer as the values cannot diverge. But
may be slower in compute.  This should be fine, importing is not
a commonly performed task.

Also fully enable pretty names, so pretty names are read and
written to the scenes.yaml file.  Also, any value matching a
default value will be removed from the scenes.yaml file for
easier human parsing.