Remove File Diff; Gracefully Handle No File / Not Defined File
Since the scenes file is a seperate file, the damage from a loss
is much lower.  Plus this seems much more reliable now and thisi
isn't something that will be frequently used, so users can make
their own backups.

Don't crash if the scene file doesn't exist or is not defined.
Will only display an error message if not defined and unable to
save scene data.