Don't Remove Group From Link Data as Default; Init Group Earlier
Fixes some esoteric bugs.

Link_defaults for a multigroup device will report all groups as
being default since we feed in the group number to generate the
default.  This fixes it, so that only group 0 or 1 devices are
treated as default.

Second, init the group value earlier, as it can affect the
link_data names returned.  Such as for a fanlinc where there is
a ramp_rate for the light, but not for the fan.