Remove Log Concept form CommandSeq; Rework Sync to Handling Logging Better
Adding Log messages to the stack on CommandSeq was a bad idea.  Any error
in executing a function that threw a TypeError would cause the function
to get exectuted again.

In order to keep sync messages displayed near the approriate events
needed to wrap the add and delete calls in functions that write the
log messages.

Also fixed a glitch wherein a DB refresh was asked for, but the
DB diff was calculated prior to receiving the refresed db data.