Process CleanupAck Messages in Response to Modem Scene Commands
Currently we just ignore these messages and assume the resulting
state of devices based on the ultimate AllLinkStatus message.

This is wrong for 2 reasons:

1. If there are a lot of devices in a scene, the length of time
that elapses from initiating the command until when the AllLinkStatus
message is received may cause a timeout.  By handling and returning
Msg.Continue, we tell the handler to stay alive and that the
command hasn't timed out.

2. If the AllLinkStatus is NAK, it just means that one or more
devices failed to send a CleanupAck.  So if we don't handle
the cleanupAcks, we erroneously assume that all devices in the
scene failed and we update none of them.  By looking at the Acks
we can update the correct ones appropriately.