Fix IOLinc Set Flags; Store Flags in DB; Track Relay and Sensor
Significant improvements to the IOLinc:

- Tracks the state of both the sensor and the relay.
- Changes the MQTT Payload to contain the sensor and relay states
by default.
- Enables the writing of flags to the device.  This design matches
how it is defined in the Insteon Documentation and also matches
how the Insteon App handles this.  The old code did not set any
flags for me.  I would be surprised if it worked on any device.
- Tracks the flags in persistent storage.  This enables things
like momentary changes and linked to relay to be enabled by
future work.