Track Relay State Internally Using Values from Device Flags
Turns off the relay after momentary_secs if momentary mode is
enabled.  There is a slight delay, as the TimedCall feature
is not a true asynch system, but it is very close and I think
within reasonable expectations.

Only turns on the relay if the proper Momentary conditions are
met.

Turn on relay if relay_linked is set and sensor is activated.

I believe this completes the modeling of the IOLinc in software.
With the exception of the "Set Button" all interactions accurately
track the IOLinc.