Remove Scene Option from IOLinc
I don't think hijacking the simulated scene call and translating
it into a modem scene was a good idea.

- It required creating a scene that would get picked up by
import_scenes and could confuse users.
- It was unnecessary for many use cases.
- It still didn't properly reflect all use cases as it didn't
allow the user to specify the on_level for D1 and so therefore
Momentary_A and Momemtary_C uses may not have been correct.

I think it is better to redirect users to create a proper scene
for their device and to provide them with the documentation on
how to do that.