Initial Draft of Hub Support as a Polling Device
This works, but not well.  The select call in start.py blocks for
3 seconds if the mqtt link doesn't have data.  This means that
polling links are very slow.

Not sure if the best option is to build something that select can
bind to?  Like a TCP socket?  However then we basically have 2
clients, the Hub client and the socket client, plus the HubClient
server.  Also not super thrilled about exposing this via TCP.