Reduce Loop Timeout for Hub; Add Message Types Seen on Hub
Reduce select/poll blocking timeout from 3 seconds to .5 seconds
when the Hub is used.  This enables basically realtime messaging
with average of .25 second delay from receipt by Hub to delivery.

Timeout remains 3 second if use_hub is not true.

This change doesn't seem to alter CPU usage, even on my relatively
minimally powered Raspberry Pi 2.

At two message types that have so far only been seen on the Hub.

027f - No idea what this is.  It is seen every 30 minutes in pairs
of two.  It could be some sort of heartbeat??  I am unable to see
any difference when the command is run, but it falls in the range
previously defined for Modem RF commands.  Simply added the msg
type, but there is no handler for it, so Protocol will parse it
but ignore it.

025C - This is more interesting.  This message looks very similar
to a Std msg received.  However, it is sent by the Hub modem when
a message cannot be delivered.  I named this an Unreachable
message.  This is not documented anywhere, nor do I see anyone
else talking about it.  It has somewhat limited use for us we
already know that a message was not Ack'd because, we never see
an Ack.  Maybe this arrives slightly sooner than we would make
that call??  Anyways, I simply added the msg type here as well.
It will get parsed out and show up in the logs, but there is
no handler for it.  No idea if newer or future PLMs will include
this command as well.