Add configuration key for MQTT client ID
The `id` key in the `mqtt` section of the configuration can be used to
supply a unique ID for each service instance.

When multiple insteon-mqtt services are connected to the same MQTT broker,
they could be configured to use separate topics, but not separate client IDs.
This resulted in the broker breaking the connections as the services connected
and reconnected.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>