No Retry on Device DB Get But Increase Timeout
If the dumping of the database times out halfway through, there
is nothing we can do but wait.  Resending the request half way
through doesn't cause the device to do anything and only risks
talking over the device messages.

Instead, since this routine is somewhat delicate, we give it
double the time_out to allow any messages to arrive.