Check Remote Battery at Interval; Signal if Below Threshold
This will check the the battery voltage of compatible remotes
at the most every 4 days.  The length is a constant that could be
changed easily.  4 days seemed reasonable.

The battery is deemed to be low if it is below 3.4v.  This is also
a constant that can easily be changed.

This uses the same low_battery topic from Battery Sensor