Make mqtt.BatterySensor Parent of Leak and Remote
Inherit mqtt code just like we did for device code.

Promote the heartbeat topic to the BatterySensor class, I think
there are other battery devices that use heartbeats.  Plus we
had halfway already included it in the BaseSensor code.

Left the ability to override topics and value templates in the
Leak and Remote just like in the Motion.  I think that this could
be helpful for someone in the future.