When Adding Last Entry Set it to All 0x00
The notes in the device db contain the following:

 The last entry in the database.  Devices show an unused, null (all
 zeros) marked with the LAST bit set in the db.  From the docs this
 shouldn't be required - the LAST bit can be a usable entry but it
 doesn't appear to work.

Which suggests that having the last entry be a usable one doesn't
work.  But some of the errors in #231 suggest that maybe
duplicating entries no longer works.

This instead sets the last entry to all 0x00s which is what a
factory reset entry looks like.  I have hopes that this may
solve generic NAK errors on new devices when adding entries using
a new address.