Don't Treat PLM NAK as Error; Fix Duplicat Logging
PLM NAK should be reported as a warning otherwise the command
line UI will report it as a fail when it isn't.

PLM NAK will likely just lead to a message timeout and retry.

Also fix some other logging that was incorrectly reporting an
ACK when a NAK had arrived.