Perform Modem DB Delete without Reliance on DB Cache
There is no reason we need the cache to perform an accurate DB
Delete.  It would be a much better user experience if we recoverd
from any errors in the delete.

To do this, we search for all entries that match addr and group
on the modem and remove them from the cache.  We then ask the
Modem to provide us only the entries that match addr and group
and add them back to the end of our db cache.  We then perform
the delete as we did previously, which should never result in a
NAK.