Enable Modify Modem DB to Work without an Accurate DB Cache
If an add is attempted and the entry already exists we get a nack
and then we attempt the same entry as an update.  The vice-versa
is true of a update as well.

This means users will no longer get the modem db is out of date
please refresh warning.