Add Get_Battery_Voltage and Set_Low_Battery_Voltage to Motion
The voltage of a 2842 motion detector battery can be requested.

This adds much of the same functionality that is present in for
Remotes.

This adds one additional feature set_low_battery_voltage which
allows the user to set the low voltage threshold.  This is needed
because the motion sensors allow users to replace the batteries
and using recharable Li-Ion batteries works with the device, but
the standard low voltage group on the device does not work.