Change Next Write Time Passed to Serial or Hub to be Dynamic
The write time was previously a timestamp that became fixed as
soon as the protocol wrote it to the Serial or Hub object.  The
Serial or Hub object would then sit on the message until that
time.

This design overlooked the possibility that the next write time
could change after the protcol sent the message to the Serial or
Hub.  This can happen if another message arrives, and the wait
time needs to be extended, or if a Link Cleanup report arrives
which could decrease the wait time.  This likely didn't cause
many errors, but wasn't quite right.

Now the write time is passed as a function which returns the
dynamic write time, so that the Serial and Hub ge the most up to
date calculation.