Move IOLinc Over to Set_State
This required a bit of hackery.  The MQTT class now makes a
direct call back to the device class to get the state of either
the relay or the sensor.  This isn't very elegant, but it is
necessary because the state output of the IOlinc produces
both the sensor and relay at the same time.  This was a mistake.