Move Manual Control to Class; Rearragne DimmerFuncs into Meta
The manual functions are split across to groups, controllers
and responders.  The base manual responders don't do anything
in response to manual commands, dimmers are the only known
devs which do anything.

Manual controllers appear to be just about anything with a button
even switches can emit manual commands.  So this needed a new
class.

DimmerFuncs was reorganized into a Meta class because the
inheritence order of ManualCtrl, Responder and Base matters
for dimmers.