Refactor MQTT client configuration to ensure consistency
When an explicit MQTT client ID is configured, the previously-created
MQTT client gets reinitialised so that it will be used. This results
in the previously-configured callback functions being lost.

A new setup_client function has been added so that both code paths will
configure the MQTT client the same way.