Simply Modem DB Delete by Using DB_Flags
Contrary to the Insteon Specification, the DB Flags can be used
in the Delete command.  While using 0x00 for the flags worked
in most circumstances, it appears that in some cases for an
unknown reason the Delete command would fail on USB Dongle style
PLMs.  It is possible that the command was also failing on other
PLM styles as well and was not noticed or reported.

Using the DB_Flags is a net win, 1) it simplifies the delete
functionality significantly and 2) it seems to solve the problem
wherein entries could not be deleted.

This hopefully fixes #351.