Rename Config.yaml -> Config-Example.yaml
Fixes #362

Home Assistant now supports defining the supervisor config in
yaml form.  Which is causing it to read our config.yaml file
as a Supervisor config and throwing all sorts of warnings.

This renames the included config file, which also has the benefit
of making it more clear to users that they need to modify the
file before using it.