Move Paho Client Log Messages into Log at Same Level
Logging of the Paho Client was previously available through a
somewhat hidden log level of 5.  This means that Paho Error
messages were only visible in the log if the logging level was
at 5 or lower.

This change moves all Paho Client log messages into the log at
their defined level.  So if a user has Error messages enabled
Paho Error messages will appear in the log.

The one exception, is that the Paho Client logs ping messages at
the debug level, which is a little verbose I think for our needs.
So all logging of ping messages have been moved to level 5.

Added notes to the config example about the level 5 logging
level.