Fix Cyclic Import Error
This is Option #2

This fixes a pylint error that was introduced when I added the
version feature.  It caused a Cyclic Import error because I was
re-importing the base insteon_mqtt module over again.  While
things worked, I agree with pylint that this is a dangerous
design.

I am not sure what the proper solution is.  I would like to access
the __version__ variable in the base module from other classes
that do not inherit from this base module.

This solution moves __version__ into its own module that does
not import anything.  This allows us to import this module into
any other module without causing a cyclic import.

This is my prefered solutio, as long as it doesn't break anything.

Error introduced in: 9e9f7061c47b2174bcd0e9e8cbee21aa49d47eaa