Change Variable Name For Comprehension; Pass Dict Copy to Publish
entries was a very generic name , disc_templates does a better
job describing what it is.

Pass a copy of a dict.  Passing the actual dict seemed to work
fine, but on the off chance that publish somehow happens async
need to pass a copy since we alter data immediately after passing
it.