Limit Characters in Discovery Topic; Revert Remove Period Address
The HA implementation of the discovery protocol only allows a
very small subset of characters for use in the node_id and
the object_id.  This replaces all un-allowed characters with an
underscore.

Also removed the ability to set the topic via yaml.  This never
really worked anyways and would create potential problems.

Reverted changing the periods in the address to underscores,
periods are still fine in the unique_id, just not in the topic
string.