Validate Yaml Files on Startup; Add Validation Schemas
This adds Cerberus as a required package.

Cerberus is used to validate the configuration yaml against an
included schema.  It will also validate the the scenes yaml
against an included schema if scenes are enabled.

The validation is done first thing on startup, and if it fails
it will print a nice clear message to the user and exit the
program.

I realize that exiting is a harsh result, but I strongly
believe that it is better for users this way.  Otherwise
configuration errors may become latent errors that are hard
to diagnose.

The validation is moderately permissive.  That is, I have
allowed unkown keys to be present in a few places.  Notably, I
did this in the devices sections of the mqtt portion.  This is
because in the past we had named a few of the topic keys
differently or allowed their use in places where they are no
longer used.

I believe, for 99.9% of users, if your installation is working
currently.  The validation will pass.  However, there may be some
odd situations that are being tolerated by the code that may
cause a validation error.