Only Change KPL is_on for load; Use is_on for LED Bits
Written originally for #433, but it doesn't look like this is
the problem.  Unit tests of the old and new code produce the
same results.

Never the less, using is_on rather than level seems more clear
for setting the led bits.

Added unit tests to catch if led bit setting is not happening.