Add Tests to Catch Error in Leak Topic; Change Default Leak Topic
The tests added will catch the error reported in #442.

However, in the process of writing these, a large complexity was
discovered.

The Leak sensor overrides the default state_topic and
state_payload configuration variables with wet_dry_topic and
wet_dry_payload.

The issue arises when the leak config does not have these values.
In that case, the default goes back to Battery_Sensor and the
topic and payload variables will change back to what is defined
in the battery_sensor.

Because the default variable names are defined in load_config
there is no way to pass the default values down from leak without
refactoring load_config across all devices.