Use Remote Group When Adding Resp Link on Modem
The group on responder links is the "remote_group".  This was correct
in the devices/Base.py but was somehow not updated in Modem.py.

Also fix bug with link data on Modem _db_update.  It is a bit
confusing, but when calling _db_add_### at the end of the function
the call is to the remote device, so the local_data is actually
remote_data.

Copy Base Device DB_Update Tests into Test Modem.
Add tests for multigroup responder links on Modem.