Remove Battery Devices from *_all Commands
If they are present, the sequence for the *_all command will
halt when it gets to a battery device and only resume when that
device has woken up and the command was sent.

This is at best confusing to the user, because they expect to
see the sequence run to completion and receive a message like
"Refresh All Devices Complete"

At worst, this is problematic if a device after a battery device
in the sequence is not getting the proper command and the user
assumes that it is.