Add MQTT Discovery Entities for Config Buttons in Home Assistant
These allow users to use the GUI in Home Assistant to run some
simple commands for devices.  These are mostly the commands needed
to install a new device.  They greatly improve the user experience.

One bummer, is that the log sensor is a pretty rudimentary way of
conveying information and status.