Fix Errors in Migration Script
Fixed some typos

The storage folder (data) is now defined in default installations
without any quoting which is how we handle the log file location.

Previously we used single quotes for storage, but the visual
code addon in Home Assistant would automatically change these to
double quotes, which is annoying.

So our sed mathing looks for any type of quoting and replaces
the default data directory with the new one without quotes.